BUBBLE SORT

+++ 0-bubble_sort.c
@@ -11,7 +11,7 @@ void	bubble_sort(int *array, size_t size)
 {
 	size_t	i;
 	size_t	j;
-	size_t	temp;
+	int	temp;
 	int	swapped;
 
 	if (!array || size < 2)
